Refactor: extract parts out since we've grown
