+++ lib/logstash/devutils/rspec/logstash_helpers.rb
@@ -1,82 +1,9 @@
 require "logstash/agent"
-require "logstash/pipeline"
 require "logstash/event"
+require "logstash/test_pipeline"
+
 require "stud/try"
 require "rspec/expectations"
-
-class LogStash::JavaPipeline
-
-  unless instance_methods(false).include?(:worker_loop)
-
-    # NOTE: copy-pasta from LS - to define the #worker_loop hook the TestPipeline relies upon.
-    def start_workers
-      @worker_threads.clear # In case we're restarting the pipeline
-      @outputs_registered.make_false
-      begin
-        maybe_setup_out_plugins
-
-        pipeline_workers = safe_pipeline_worker_count
-        batch_size = settings.get("pipeline.batch.size")
-        batch_delay = settings.get("pipeline.batch.delay")
-
-        max_inflight = batch_size * pipeline_workers
-
-        config_metric = metric.namespace([:stats, :pipelines, pipeline_id.to_s.to_sym, :config])
-        config_metric.gauge(:workers, pipeline_workers)
-        config_metric.gauge(:batch_size, batch_size)
-        config_metric.gauge(:batch_delay, batch_delay)
-        config_metric.gauge(:config_reload_automatic, settings.get("config.reload.automatic"))
-        config_metric.gauge(:config_reload_interval, settings.get("config.reload.interval"))
-        config_metric.gauge(:dead_letter_queue_enabled, dlq_enabled?)
-        config_metric.gauge(:dead_letter_queue_path, dlq_writer.get_path.to_absolute_path.to_s) if dlq_enabled?
-        config_metric.gauge(:ephemeral_id, ephemeral_id)
-        config_metric.gauge(:hash, lir.unique_hash)
-        config_metric.gauge(:graph, ::LogStash::Config::LIRSerializer.serialize(lir))
-        config_metric.gauge(:cluster_uuids, resolve_cluster_uuids)
-
-        pipeline_log_params = default_logging_keys(
-            "pipeline.workers" => pipeline_workers,
-            "pipeline.batch.size" => batch_size,
-            "pipeline.batch.delay" => batch_delay,
-            "pipeline.max_inflight" => max_inflight,
-            "pipeline.sources" => pipeline_source_details)
-        @logger.info("Starting pipeline", pipeline_log_params)
-
-        if max_inflight > MAX_INFLIGHT_WARN_THRESHOLD
-          @logger.warn("CAUTION: Recommended inflight events max exceeded! Logstash will run with up to #{max_inflight} events in memory in your current configuration. If your message sizes are large this may cause instability with the default heap size. Please consider setting a non-standard heap size, changing the batch size (currently #{batch_size}), or changing the number of pipeline workers (currently #{pipeline_workers})", default_logging_keys)
-        end
-
-        filter_queue_client.set_batch_dimensions(batch_size, batch_delay)
-
-        pipeline_workers.times do |t|
-          thread = Thread.new do
-            Util.set_thread_name("[#{pipeline_id}]>worker#{t}")
-            ThreadContext.put("pipeline.id", pipeline_id)
-            worker_loop #
-          end
-          @worker_threads << thread
-        end
-
-        begin
-          start_inputs
-        rescue => e
-          shutdown_workers
-          raise e
-        end
-      ensure
-        @ready.make_true
-      end
-    end
-
-    java_import org.logstash.execution.WorkerLoop
-
-    def worker_loop
-      WorkerLoop.new(lir_execution, filter_queue_client, @events_filtered, @events_consumed,
-                     @flushRequested, @flushing, @shutdownRequested, @drain_queue).run
-    end
-  end
-
-end
 
 require "logstash/environment"
 
@@ -103,117 +30,7 @@ module LogStashHelper
     @@excluded_tags
   end
 
-  class TestPipeline < LogStash::JavaPipeline
-    public :flush_filters
-
-    attr_reader :test_read_client
-
-    def run_with(events)
-      if inputs&.any? # will work but might be unintended
-        config = "\n #{config_str}" if $VERBOSE
-        warn "#{self} pipeline is getting events pushed manually while having inputs: #{inputs.inspect}  #{config}"
-      end
-      # TODO could we handle an generator (Enumerator) ?
-      queue.write_client.push_batch events.to_a
-      @test_read_client = EventTrackingQueueReadClientDelegator.new filter_queue_client
-      run
-    end
-
-    # @override
-    def worker_loop
-      read_client = @test_read_client || filter_queue_client
-      WorkerLoop.new(lir_execution, read_client, @events_filtered, @events_consumed,
-                     @flushRequested, @flushing, @shutdownRequested, @drain_queue).run
-
-      super
-    end
-
-    class EventTrackingQueueReadClientDelegator
-      include org.logstash.execution.QueueReadClient
-      java_import org.logstash.execution.QueueReadClient
-
-      attr_reader :processed_events
-
-      def initialize(delegate)
-        # NOTE: can not use LogStash::MemoryReadClient#read_batch due its JavaObject wrapping
-        @delegate = delegate.to_java(QueueReadClient)
-        @processed_events = []
-      end
-
-      # @override QueueBatch readBatch() throws InterruptedException;
-      def readBatch
-        QueueBatchDelegator.new(self, @delegate.read_batch)
-      end
-
-      # @override void closeBatch(QueueBatch batch) throws IOException;
-      def closeBatch(batch)
-        @delegate.close_batch(batch)
-      end
-
-      # @override boolean isEmpty();
-      def isEmpty
-        @delegate.empty?
-      end
-
-      # @override QueueBatch newBatch();
-      def newBatch
-        @delegate.new_batch
-      end
-
-      # @override void startMetrics(QueueBatch batch);
-      def startMetrics(batch)
-        @delegate.start_metrics(batch)
-      end
-
-      # @override void addOutputMetrics(int filteredSize);
-      def addOutputMetrics(filteredSize)
-        @delegate.add_output_metrics(filteredSize)
-      end
-
-      # @override void addFilteredMetrics(int filteredSize);
-      def addFilteredMetrics(filteredSize)
-        @delegate.add_filtered_metrics(filteredSize)
-      end
-
-      def filtered_events(events)
-        @processed_events.concat(events)
-      end
-
-    end
-
-    class QueueBatchDelegator
-      include org.logstash.execution.QueueBatch
-
-      def initialize(event_tracker, delegate)
-        @event_tracker = event_tracker
-        @delegate = delegate
-      end
-
-      # @override RubyArray to_a();
-      def to_a
-        @delegate.to_a.tap do |events|
-          # filters out rogue (cancelled) events
-          @event_tracker.filtered_events events
-        end
-      end
-
-      # @override int filteredSize();
-      def filteredSize
-        @delegate.to_java.filtered_size
-      end
-
-      # @override void merge(IRubyObject event);
-      def merge(event)
-        @delegate.merge(event)
-      end
-
-      # @override void close() throws IOException;
-      def close
-        @delegate.close
-      end
-
-    end
-  end
+  TestPipeline = LogStash::TestPipeline
 
   DEFAULT_NUMBER_OF_TRY = 5
   DEFAULT_EXCEPTIONS_FOR_TRY = [RSpec::Expectations::ExpectationNotMetError]
